requirements.txt
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google==3.0.0
google-analytics-data==0.16.1
google-api-core==2.11.0
google-auth==2.16.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
